Added SCCDemote transformation
